# PR: add new cached procedures to compute the tensor, which avoid redundent computation in the for loops.