Dear Giovanni,

I created some new functions to compute tensors, which avoid some redundant computation in the for loop.

I have tested the performance of the modification, which will speed up the computation in an order close to **Csize**. The following is a description of my test case.

1. Problem: a modification based on **06POD_RBF**
2. Number of cells: 198000
3. Three tests are performed, and the results are listed in the table.

| Num modes | Original | Modified | Speedup ratio |
| --------- | --------     | -------- | ------------- |
| 3          | 19.58     | 4.62    | 4.24          |
| 5          | 93.33     | 14.71  | 6.34          |
| 10        | 691.85   | 60.6    | 11.42        |

A similar modification is applied to
1. divMomentum
2. turbulenceTensor1, turbulenceTensor2
3. turbulencePPETensor1, turbulencePPETensor2

The attached file contains a test case for comparing the Tensors generated by the new function with those from the original one, both in serial and parallel execution.
[06POD_RBF_cached_test.tar.gz](https://github.com/user-attachments/files/20525106/06POD_RBF_cached_test.tar.gz)

With best regards,
Shenhui

The explanation is listed below:

The computation of the [**convective_term_tens**](https://github.com/ITHACA-FV/ITHACA-FV/blob/ff7b59ff50d9078d5033c219f5c9615455f81f0a/src/ITHACA_FOMPROBLEMS/steadyNS/steadyNS.C#L997) contains three-layer loops. Therefore, the part  `fvc::div(linearInterpolate(L_U_SUPmodes[j]) & L_U_SUPmodes[j].mesh().Sf(), L_U_SUPmodes[k])` is computed **i** times, which is very costly.

```
for (label i = 0; i < Csize; i++)
{
    for (label j = 0; j < Csize; j++)
    {
        for (label k = 0; k < Csize; k++)
        {
            if (fluxMethod == "consistent")
            {
                C_tensor(i, j, k) = fvc::domainIntegrate(L_U_SUPmodes[i] & fvc::div(
                                        L_PHImodes[j],
                                        L_U_SUPmodes[k])).value();
            }
            else
            {
                C_tensor(i, j, k) = fvc::domainIntegrate(L_U_SUPmodes[i] & fvc::div(
                                        linearInterpolate(L_U_SUPmodes[j]) & L_U_SUPmodes[j].mesh().Sf(),
                                        L_U_SUPmodes[k])).value();
            }
        }
    }
```

The above code is OK for small-scale problems. But I am applying it to problems with several million cells, and the computation of the tensor becomes the bottleneck of the projection procedure.

I suggest computing the intermediate PrtList to accelerate the computation as follows:

```
PtrList< PtrList<volVectorField> > divCache(Csize);
for (label j = 0; j < Csize; ++j)
{
    PtrList<volVectorField> innerCache(Csize);

    for (label k = 0; k < Csize; ++k)
    {
        volVectorField* divFieldPtr = nullptr;

        if (fluxMethod == "consistent")
        {
            divFieldPtr = new volVectorField
            (
                fvc::div(L_PHImodes[j], L_U_SUPmodes[k])
            );
        }
        else
        {
            surfaceScalarField interp
            (
                linearInterpolate(L_U_SUPmodes[j]) & L_U_SUPmodes[j].mesh().Sf()
            );

            divFieldPtr = new volVectorField
            (
                fvc::div(interp, L_U_SUPmodes[k])
            );
        }
        innerCache.set(k, divFieldPtr);
    }
    divCache.set(j, new PtrList<volVectorField>(innerCache));
}

for (label i = 0; i < Csize; i++)
{
    for (label j = 0; j < Csize; j++)
    {
        for (label k = 0; k < Csize; k++)
        {
            C_tensor(i, j, k) = fvc::domainIntegrate(L_U_SUPmodes[i] & divCache[j][k]).value();
        }
    }
}
```